refactor(yahoo): move YahooStockPriceProvider to HostedService
YahooStockPriceProvider is a strategy adapter for IStockPriceProvider,
not a BaseRepository<T> — having it in Equibles.Yahoo.Repositories was
a category error that forced Worker.Host to register it manually.

Now lives in Equibles.Yahoo.HostedService.Services and is registered as
part of AddYahooWorker(), so any host composing the Yahoo worker gets
price-provider wiring for free. Drops the ad-hoc cross-module DI line
from Worker.Host/Program.cs.
