+++ src/Equibles.Yahoo.HostedService/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Equibles.Core.AutoWiring;
+using Equibles.Core.Contracts;
 using Equibles.Yahoo.HostedService.Services;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -8,6 +9,9 @@ public static class ServiceCollectionExtensions {
     public static IServiceCollection AddYahooWorker(this IServiceCollection services) {
         services.AutoWireServicesFrom<YahooPriceImportService>();
         services.AutoWireServicesFrom<Equibles.Integrations.Yahoo.YahooFinanceClient>();
+
+        // Yahoo is the OSS source of stock prices for Holdings valuation.
+        services.AddScoped<IStockPriceProvider, YahooStockPriceProvider>();
 
         services.AddHostedService<YahooPriceScraperWorker>();
